add alert dialog when deleting product

+++ src/app/(admin)/products/ProductActions.tsx
@@ -1,3 +1,16 @@
+"use client";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
@@ -8,6 +21,7 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { Product } from "@prisma/client";
 import { Ellipsis } from "lucide-react";
+import Link from "next/link";
 import { toast } from "sonner";
 import { deleteProductAction, updateProductStatusAction } from "./actions";
 
@@ -22,7 +36,11 @@ export function ProductActions({ product }: { product: Product }) {
       </DropdownMenuTrigger>
 
       <DropdownMenuContent align="end">
-        {/* <DropdownMenuItem>Edit</DropdownMenuItem> */}
+        <DropdownMenuItem asChild>
+          <Link href={`/products/${product.id}/edit`} className="cursor-pointer font-medium">
+            Edit
+          </Link>
+        </DropdownMenuItem>
 
         <ActivateDropdownItem productId={product.id} />
 
@@ -123,19 +141,40 @@ function ArchiveDropdownItem({ productId }: { productId: string }) {
 function DeleteDropdownItem({ productId }: { productId: string }) {
   return (
     <DropdownMenuItem asChild>
-      <Button
-        onClick={async () => {
-          toast.promise(deleteProductAction(productId), {
-            loading: "Deleting product...",
-            success: "Product deleted",
-            error: (error) => error.message,
-          });
-        }}
-        className="w-full cursor-pointer justify-start text-destructive focus-visible:text-destructive focus-visible:ring-0"
-        variant={"ghost"}
-      >
-        Delete
-      </Button>
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button
+            variant={"ghost"}
+            className="w-full cursor-pointer justify-start px-2 py-1 text-destructive hover:text-destructive focus-visible:text-destructive focus-visible:ring-0"
+          >
+            Delete
+          </Button>
+        </AlertDialogTrigger>
+
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete the product.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              variant={`destructive`}
+              onClick={async () => {
+                toast.promise(deleteProductAction(productId), {
+                  loading: "Deleting product...",
+                  success: "Product deleted",
+                  error: (error) => error.message,
+                });
+              }}
+            >
+              Yes, delete product
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </DropdownMenuItem>
   );
 }
